// src/models/ClubLocation.ts
export type FfrLeagueId =
  | "auvergne_rhone_alpes"
  | "bourgogne_franche_comte"
  | "bretagne"
  | "centre_val_de_loire"
  | "corse"
  | "grand_est"
  | "hauts_de_france"
  | "normandie"
  | "nouvelle_aquitaine"
  | "occitanie"
  | "pays_de_la_loire"
  | "provence_alpes_cote_d_azur"
  | "ile_de_france";